Moved Mike's new classes to the correct package for this version of bluedb
