+++ BlueDBOnDisk/src/test/java/org/bluedb/disk/CachedSingleThreadingPoolTest.java
@@ -1,10 +1,11 @@
-package io.bluedb.util;
+package org.bluedb.disk;
 
 import static org.junit.Assert.*;
 
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.Future;
 
+import org.bluedb.disk.CachedSingleThreadingPool;
 import org.junit.Test;
 
 public class CachedSingleThreadingPoolTest {
